+++ kitty/options/to-c-generated.h
@@ -981,19 +981,6 @@ convert_from_opts_tab_bar_style(PyObject *py_opts, Options *opts) {
 }
 
 static void
-convert_from_python_tab_bar_min_tabs(PyObject *val, Options *opts) {
-    opts->tab_bar_min_tabs = PyLong_AsUnsignedLong(val);
-}
-
-static void
-convert_from_opts_tab_bar_min_tabs(PyObject *py_opts, Options *opts) {
-    PyObject *ret = PyObject_GetAttrString(py_opts, "tab_bar_min_tabs");
-    if (ret == NULL) return;
-    convert_from_python_tab_bar_min_tabs(ret, opts);
-    Py_DECREF(ret);
-}
-
-static void
 convert_from_python_tab_bar_background(PyObject *val, Options *opts) {
     opts->tab_bar_background = color_or_none_as_int(val);
 }
@@ -1508,8 +1495,6 @@ convert_opts_from_python_opts(PyObject *py_opts, Options *opts) {
     convert_from_opts_tab_bar_margin_height(py_opts, opts);
     if (PyErr_Occurred()) return false;
     convert_from_opts_tab_bar_style(py_opts, opts);
-    if (PyErr_Occurred()) return false;
-    convert_from_opts_tab_bar_min_tabs(py_opts, opts);
     if (PyErr_Occurred()) return false;
     convert_from_opts_tab_bar_background(py_opts, opts);
     if (PyErr_Occurred()) return false;